flaskServer/requirements.txt
blinker
cairocffi
certifi
cffi
charset-normalizer
click
click-default-group
cloup
colorama
colour==0.1.5
commonmark
decorator
distlib==0.3.8
filelock==3.14.0
Flask==3.0.3
fsspec==2024.3.1
gensim==4.3.2
Gunicorn
glcontext==2.3.7
idna
imageio
imageio-ffmpeg
intel-openmp==2021.4.0
isosurfaces==0.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
llvmlite
mapbox-earcut==1.0.1
MarkupSafe==2.1.5
mkl==2021.4.0
moderngl==5.7.4
moderngl-window==2.4.2
more-itertools==10.2.0
moviepy
mpmath==1.3.0
multipledispatch==0.6.0
mysql-connector-python==8.0.32
networkx==2.8.8
numba==0.59.1
numpy==1.24.1
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python==4.7.0.68
pangocairocffi==0.4.0
pangocffi==0.8.0
Pillow
platformdirs==4.2.1
proglog==0.1.10
progressbar==2.5
pycparser==2.21
pydub==0.25.1
pyglet==2.0.3
Pygments==2.14.0
pyreadline==2.1
pyrr==0.10.3
pytube
regex==2024.4.28
requests==2.28.1
rich
scikit-learn==1.4.2
scipy==1.10.0
screeninfo==0.8.1
six==1.16.0
skia-pathops==0.7.4
smart-open==7.0.4
srt==3.5.2
svgelements==1.9.0
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.5.0
tiktoken==0.6.0
torch==2.3.0
tqdm==4.64.1
typing_extensions
urllib3
virtualenv==20.26.1
watchdog==4.0.0
Werkzeug==3.0.3
wrapt==1.16.0
